refactor the structure of the moe folder, involving files such as layer, policy, unit test.
